# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AMI-FieldWorkManager API",
    "version": "1.0.0",
    "description": ""
  },
  "paths": {
...
    "/api/solicitudes": {
      "get": {
        "operationId": "apps_operations_api_listar_solicitudes",
        "summary": "Listar Solicitudes",
        "parameters": [
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Estado",
              "type": "array"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "equipo",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipo"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "urgencia",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Urgencia"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "desde",
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "hasta",
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasta"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "mias",
            "schema": {
              "default": false,
              "title": "Mias",
              "type": "boolean"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page_size",
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedSolicitudOut"
                }
              }
            }
          }
        },
        "description": "`desde`/`hasta` filtran por fecha programada; `mias` por creador.",
        "tags": [
          "operaciones"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ]
      },
      "post": {
        "operationId": "apps_operations_api_crear_solicitud",
        "summary": "Crear Solicitud",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitudOut"
                }
              }
            }
          }
        },
        "tags": [
          "operaciones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudIn"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    },
    "/api/solicitudes/{solicitud_id}": {
      "get": {
        "operationId": "apps_operations_api_obtener_solicitud",
        "summary": "Obtener Solicitud",
        "parameters": [
          {
            "in": "path",
            "name": "solicitud_id",
            "schema": {
              "title": "Solicitud Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitudOut"
                }
              }
            }
          }
        },
        "tags": [
          "operaciones"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ]
      },
      "patch": {
        "operationId": "apps_operations_api_editar_solicitud",
        "summary": "Editar Solicitud",
        "parameters": [
          {
            "in": "path",
            "name": "solicitud_id",
            "schema": {
              "title": "Solicitud Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitudOut"
                }
              }
            }
          }
        },
        "tags": [
          "operaciones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudPatch"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "SessionAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "apps_operations_api_borrar_solicitud",
        "summary": "Borrar Solicitud",
        "parameters": [
          {
            "in": "path",
            "name": "solicitud_id",
            "schema": {
              "title": "Solicitud Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "description": "Falla con 409 si la solicitud ya forma parte de una orden de trabajo.",
        "tags": [
          "operaciones"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    },
    "/api/solicitudes/{solicitud_id}/cancelar": {
      "post": {
        "operationId": "apps_operations_api_cancelar_solicitud",
        "summary": "Cancelar Solicitud",
        "parameters": [
          {
            "in": "path",
            "name": "solicitud_id",
            "schema": {
              "title": "Solicitud Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitudOut"
                }
              }
            }
          }
        },
        "tags": [
          "operaciones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MotivoIn"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    },
    "/api/unidades-campo": {
      "get": {
        "operationId": "apps_operations_api_listar_unidades",
        "summary": "Listar Unidades",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UnidadCampoSchema"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": [
          "operaciones"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ]
      },
      "put": {
        "operationId": "apps_operations_api_guardar_unidades",
        "summary": "Guardar Unidades",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UnidadCampoSchema"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            }
          }
        },
        "description": "Reemplaza todas las composiciones por defecto de las cuadrillas.",
        "tags": [
          "operaciones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/UnidadCampoSchema"
                },
                "title": "Payload",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    },
    "/api/ordenes": {
      "get": {
        "operationId": "apps_operations_api_listar_ordenes",
        "summary": "Listar Ordenes",
        "parameters": [
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Estado",
              "type": "array"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "desde",
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "hasta",
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasta"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "tecnico",
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tecnico"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page_size",
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOrdenOut"
                }
              }
            }
          }
        },
        "description": "`desde`/`hasta` filtran por fecha de creación; `tecnico` por técnico asignado.",
        "tags": [
          "operaciones"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ]
      },
      "post": {
        "operationId": "apps_operations_api_crear_ordenes",
        "summary": "Crear Ordenes",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrdenDetalleOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            }
          }
        },
        "description": "Crea una o varias OT (p. ej., todas las rutas del tablero de planificación).",
        "tags": [
          "operaciones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/OrdenIn"
                },
                "title": "Payload",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    },
    "/api/ordenes/{orden_id}": {
      "get": {
        "operationId": "apps_operations_api_obtener_orden",
        "summary": "Obtener Orden",
        "parameters": [
          {
            "in": "path",
            "name": "orden_id",
            "schema": {
              "title": "Orden Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDetalleOut"
                }
              }
            }
          }
        },
        "tags": [
          "operaciones"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "apps_operations_api_borrar_orden",
        "summary": "Borrar Orden",
        "parameters": [
          {
            "in": "path",
            "name": "orden_id",
            "schema": {
              "title": "Orden Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "operaciones"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    },
...
    "/api/trabajos/{trabajo_id}": {
      "get": {
        "operationId": "apps_operations_api_obtener_trabajo",
        "summary": "Obtener Trabajo",
        "parameters": [
          {
            "in": "path",
            "name": "trabajo_id",
            "schema": {
              "title": "Trabajo Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrabajoOut"
                }
              }
            }
          }
        },
        "tags": [
          "operaciones"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PagedSolicitudOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SolicitudOut"
            },
            "title": "Items",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "PagedSolicitudOut",
        "type": "object"
      },
...
      "MotivoIn": {
        "properties": {
          "motivo": {
            "minLength": 1,
            "title": "Motivo",
            "type": "string"
          }
        },
        "required": [
          "motivo"
        ],
        "title": "MotivoIn",
        "type": "object"
      },
      "UnidadCampoSchema": {
        "properties": {
          "vehiculo": {
            "title": "Vehiculo",
            "type": "string"
          },
          "tecnicos": {
            "items": {
              "type": "integer"
            },
            "title": "Tecnicos",
            "type": "array"
          }
        },
        "required": [
          "vehiculo",
          "tecnicos"
        ],
        "title": "UnidadCampoSchema",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      }
    }
  },
  "servers": []
}